+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ## Next
 
+- Fix counter not being reset between test cases (#12, kudos to @komkovla)
+
 ## 0.4.0
 
 - Add UIKit support for snapshot testing (#11, kudos to @komkovla)